Massive refactoring job. Merged builder package with mudkip/lib because they will both share a large amount of the builder content. They are too similar to warrant separate packages. Considering the amount of types defined in a game world, the datastore will be reworked as well. The generic object (de)serialization bits in the datastore are removed.

+++ lib/game/userinfo.go
@@ -1,9 +1,9 @@
 package lib
 
 type UserInfo struct {
-	Id         uint16
+	Id         int64
 	Name       string
 	Password   string
-	Zone       uint16
+	Zone       int64
 	Registered int64
 }
